Do "similar" changes from Guichan's changes from e82f223b2ab3cc2ea81807097d52130255836a7d to 1f9ed38079833ed9527cbcfead5c3caebeb85780 (Jan 2nd 2008).
Factorize examples code.
